This script looks for systems that do not have a body and fetches the body data from EDSM and inserts it into the bodies model.
Its quite slow because of the EDSM lookups. However we should expect after the initial load
that we wont be inserting very often.

It needs a further modification because we wont want all bodies to be loaded
only for specific models so we will need to join or union on specific models